+++ src/routes/login/+page.svelte
@@ -3,6 +3,7 @@
   import { goto } from '$app/navigation';
   import { onMount } from 'svelte';
   import axios from 'axios';
+  import Cookies from 'js-cookie';
 
   onMount(async () => {
     const loggedIn = window.localStorage.getItem('loggedIn') == 'true';
@@ -36,7 +37,7 @@
     }
     if (!noUsernameError && !noPasswordError) {
       try {
-        const response = await fetch('http://localhost:3011/login', {
+        const response = await fetch('http://localhost:3011/auth/login', {
           method: 'POST',
           headers: {
             'Content-Type': 'application/json',
@@ -49,8 +50,9 @@
         const data = await response.json();
         console.log(data);
         if (response.ok) {
-          localStorage.setItem('userToken', data.token);
-          localStorage.setItem('loggedIn', 'true');
+          const token = data.token;
+          console.log(token);
+          Cookies.set('cubySession', token);
           goto('/home');
         } else {
           if (response.status === 401 && data.error === 'Authentication failed. User not found.') {
@@ -70,7 +72,10 @@
   const googleLogin = async () => {
     try {
       const response = await axios.get('http://localhost:3011/auth/google'); // Make sure this path matches your API Gateway route
-      window.location.href = response.data.redirectUrl;
+      const token = data.token;
+      console.log(token);
+      Cookies.set('cubySession', token);
+      goto('/home');
     } catch (error) {
       console.error('Error initiating Google login:', error.message);
     }
